my-framework/src/test/java/TestCase.java
import java.lang.reflect.Method;

public abstract class TestCase {

    protected String testCaseName;

    public TestCase(String testCaseName) {
        this.testCaseName = testCaseName;
    }

    public void run() {
        before();
        runTestCase();
        after();
    }

    protected void before() {}

    public void runTestCase() {
        try {
            Method method = this.getClass().getMethod(testCaseName, null);
            method.invoke(this, null);
        } catch (Exception e) {
            throw new RuntimeException(e);
        }
    }

    protected void after() {}
}
